realigning

+++ base/include/asm-x86_64/rtai_hal.h
@@ -359,7 +359,7 @@ irqreturn_t rtai_broadcast_to_local_timers(int irq,
 #define _send_sched_ipi(dest) \
 do { \
 	apic_wait_icr_idle(); \
-	apic_write_around(APIC_ICR2, SET_APIC_DEST_FIELD(dest)); \
+	apic_write_around(APIC_ICR2, (int)SET_APIC_DEST_FIELD((unsigned int)dest)); \
 	apic_write_around(APIC_ICR, APIC_DEST_LOGICAL | SCHED_VECTOR); \
 } while (0)
